Speed dependent fov now works.
Some spikes caused by lag  in fov canm be seen. By using a 25 frame buffer these are now not as severe as they could have been. This causes delay in the speed effect though. A Log() function has also been implemented which can be used instead to minimize delay. This has big spikes though. Therefore a Avg() function is used at this time. Log() can be used when position updating is done on the client