Updated license.
Switched to BSD3 to comply with the Artemis license. Added Artemis license
and notice of that the ES project is based on Artemis.

Note: The ES project is not a straight port of Artemis, far from so. It is
however based/inpsired on Artemis. It hasn't been scrutinized and/or
sanitized from parts that may stem from Artemis by GAMADU. The Artemis
license is therefore added as a precautionary action.